requirements.txt
click
biopython
PyQt5
ete3